chore: refactored the components folder . It now has a subfolder called common
