+++ pages/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import type { NextPage } from 'next';
-import Layout from '@/components/Layout';
-import { Box, Button } from '@chakra-ui/react';
-import { signIn, useSession } from 'next-auth/client';
+import Layout from '@/components/common/Layout';
+import { Box } from '@chakra-ui/react';
+import { useSession } from 'next-auth/client';
 
 const Home: NextPage = () => {
     const [session, loading] = useSession();
